externalize utility functions as a separated module
