+++ apps/main/package.json
@@ -78,6 +78,7 @@
         "typescript": "^5.2.2",
         "webpack": "^5.88.2",
         "webpack-bundle-analyzer": "^4.9.0",
-        "webpack-cli": "^5.1.4"
+        "webpack-cli": "^5.1.4",
+        "utils": "workspace:*"
     }
 }
